pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dynamic-agent-assembler"
version = "0.1.0"
description = "Runtime Agent Team Construction with Vector-based Capability Matching"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Yoshikondo", email = "yoshikondo@example.com"}
]
keywords = ["ai", "agents", "multi-agent", "capability-matching", "vector-search"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "tenacity>=8.2.0",
    "PyYAML>=6.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
openai = ["openai>=1.3.0"]
anthropic = ["anthropic>=0.18.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py310"
